Moved GenerateDataTaskParams into the correct package.
